# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.1.1",
    "title": "Questify API docs",
    "description": "API documentation for GoIT project 'Questify'. [Backend github repository](https://github.com/Powerman-code/finalTeamProject-tasks-server)",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "servers": [{"url":"http://localhost:8080/"},{"url":"https://goit-backend-23.herokuapp.com/"}],
...
  "paths": {
...
    "/api/card": {
      "post": {
        "tags": ["Card"],
        "summary": "Create card",
        "security": [{ "Bearer": [] }],
        "parameters": [],
        "requestBody": {
          "description": "Card's object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid request body)",
            "content": {}
          },
          "404": {
            "description": "Invalid user / Invalid session",
            "content": {}
          }
        }
      
      },
     
      "get": {
        "tags": ["Card"],
        "summary": "Get all user card",
        "security": [{ "Bearer": [] }],
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (no token provided)",
            "content": {}
          },
          "404": {
            "description": "Invalid user / Invalid session",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CardRequest": {
        "required": ["title", "difficulty", "type", "date"],
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Card's title",
            "example": "Take out the trash"
          },
          "difficulty": {
            "type": "string",
            "description": "Card's difficulty",
            "enum": ["Easy", "Normal", "Hard"]
          },
          "category": {
            "type": "string",
            "description": "Card's category",
            "enum": ["Stuff", "Family", "Health", "Learning", "Leisure", "Work"]
          },
          "date": {
            "type": "string",
            "description": "Card's date (deadline, if card is challenge)",
            "example": "2020-12-31"
          },
          "time": {
            "type": "string",
            "description": "Card's time (deadline, if card is challenge)",
            "example": "20:30"
          },
          "type": {
            "type": "string",
            "description": "Card's type",
            "enum": ["Task", "Challenge"]
          }
        }
      },
      "CardResponse": {
        "type": "object",
        "properties": {
          "createdCard": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string",
                "description": "Card's title",
                "example": "Take out the trash"
              },
              "difficulty": {
                "type": "string",
                "description": "Card's difficulty",
                "enum": ["Easy", "Normal", "Hard"]
              },
              "category": {
                "type": "string",
                "description": "Card's category",
                "enum": [
                  "Stuff",
                  "Family",
                  "Health",
                  "Learning",
                  "Leisure",
                  "Work"
                ]
              },
              "date": {
                "type": "string",
                "description": "Card's date (deadline, if card is challenge)",
                "example": "2020-12-31"
              },
              "time": {
                "type": "string",
                "description": "Card's time (deadline, if card is challenge)",
                "example": "20:30"
              },
              "status": {
                "type": "string",
                "description": "Card's status",
                "enum": ["Incomplete", "Complete"]
              },
              "type": {
                "type": "string",
                "description": "Card's type",
                "enum": ["Task", "Challenge"]
              },
              "_id": {
                "type": "string",
                "description": "Card's id",
                "example": "507f1f77bcf86cd799439013"
              }
            }
          }
        }
      },
...
      "GetCardResponse": {
        "type": "object",
        "properties": {
          "cards": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "User's card",
              "properties": {
                "title": {
                  "type": "string",
                  "description": "Card's title",
                  "example": "Take out the trash"
                },
                "difficulty": {
                  "type": "string",
                  "description": "Card's difficulty",
                  "enum": ["Easy", "Normal", "Hard"]
                },
                "date": {
                  "type": "string",
                  "description": "Card's date (deadline, if card is challenge)",
                  "example": "2020-12-31"
                },
                "time": {
                  "type": "string",
                  "description": "Card's time (deadline, if card is challenge)",
                  "example": "20:30"
                },
                "status": {
                  "type": "string",
                  "description": "Card's status",
                  "enum": ["Incomplete", "Complete"]
                },
                "type": {
                  "type": "string",
                  "description": "Card's type",
                  "enum": ["Task", "Challenge"]
                },
                "_id": {
                  "type": "string",
                  "description": "Card's id",
                  "example": "507f1f77bcf86cd799439013"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}